VCV-183-1: create shared site hierarchy/site leaf rows/collapsible location group components
