+++ src/features/review/sites-list/components/sites/review-site-hierarchy.tsx
@@ -1,215 +1,87 @@
 'use client';
 
 import type { Site } from '@/api/site/validation/site-schema';
-import { isLegacySite } from '@/lib/location/location-query';
-import {
-    getLocationTypeName,
-    getSiteAndDescendants,
-} from '@/lib/location/site-tree';
-import { useMemo } from 'react';
+import SiteHierarchy from '@/features/review/shared/site-hierarchy';
+import ReviewVisitCoverageBadge from './review-visit-coverage-badge';
 import ReviewSiteLeafRows from './review-site-leaf-rows';
-import ReviewCollapsibleLocationGroup from './review-collapsible-location-group';
 import type { ReviewSiteSessionSummary } from '../../utils/review-site-session-summary';
 
-const LEGACY_HIERARCHY_LEVELS = [
-    { key: 'subCounty', label: 'Subcounty' },
-    { key: 'healthCenter', label: 'Health Center' },
-    { key: 'parish', label: 'Parish' },
-    { key: 'villageName', label: 'Village' },
-    { key: 'houseNumber', label: 'House' },
-] as const;
-
-interface LegacySiteHierarchyProps {
-    sites: Site[];
-    depth: number;
-    parentPath: string;
-    sessionCountsBySiteId: Map<number, ReviewSiteSessionSummary>;
-    expandedSitePaths: Set<string>;
-    onToggle: (path: string) => void;
-}
-
-function LegacySiteHierarchy({
-    sites,
-    depth,
-    parentPath,
-    sessionCountsBySiteId,
-    expandedSitePaths,
-    onToggle,
-}: LegacySiteHierarchyProps) {
-    const currentLevel = LEGACY_HIERARCHY_LEVELS[depth];
-    const isLeafLevel = depth === LEGACY_HIERARCHY_LEVELS.length - 1;
-
-    const sortedLocationGroups = useMemo(() => {
-        if (!currentLevel) return [];
-        const sitesByLocationName: Record<string, Site[]> = {};
-        for (const site of sites) {
-            const locationName = site[currentLevel.key] ?? 'Unknown';
-            (sitesByLocationName[locationName] ??= []).push(site);
-        }
-        return Object.entries(sitesByLocationName).sort(([a], [b]) =>
-            a.localeCompare(b),
-        );
-    }, [sites, currentLevel]);
-
-    if (!currentLevel) return null;
-
-    if (isLeafLevel) {
-        return (
-            <ReviewSiteLeafRows
-                sites={sites}
-                getDisplayName={site => site[currentLevel.key] ?? 'Unknown'}
-                sessionCountsBySiteId={sessionCountsBySiteId}
-            />
-        );
-    }
-
-    return (
-        <div className="space-y-1">
-            {sortedLocationGroups.map(([locationName, sitesInGroup]) => (
-                <ReviewCollapsibleLocationGroup
-                    key={`${parentPath}/${locationName}`}
-                    locationName={locationName}
-                    locationTypeName={currentLevel.label}
-                    sitesInGroup={sitesInGroup}
-                    parentPath={parentPath}
-                    sessionCountsBySiteId={sessionCountsBySiteId}
-                    expandedSitePaths={expandedSitePaths}
-                    onToggle={onToggle}
-                >
-                    <LegacySiteHierarchy
-                        sites={sitesInGroup}
-                        depth={depth + 1}
-                        parentPath={`${parentPath}/${locationName}`}
-                        sessionCountsBySiteId={sessionCountsBySiteId}
-                        expandedSitePaths={expandedSitePaths}
-                        onToggle={onToggle}
-                    />
-                </ReviewCollapsibleLocationGroup>
-            ))}
-        </div>
-    );
-}
-
-interface HierarchicalSiteHierarchyProps {
-    sites: Site[];
-    parentSiteId: number | undefined;
-    parentPath: string;
-    sessionCountsBySiteId: Map<number, ReviewSiteSessionSummary>;
-    expandedSitePaths: Set<string>;
-    onToggle: (path: string) => void;
-}
-
-function HierarchicalSiteHierarchy({
-    sites,
-    parentSiteId,
-    parentPath,
-    sessionCountsBySiteId,
-    expandedSitePaths,
-    onToggle,
-}: HierarchicalSiteHierarchyProps) {
-    const childSites = useMemo(
-        () =>
-            sites
-                .filter(site => site.parentId === parentSiteId)
-                .sort((a, b) => (a.name ?? '').localeCompare(b.name ?? '')),
-        [sites, parentSiteId],
-    );
-
-    const isLeafLevel = childSites.every(
-        site => !sites.some(otherSite => otherSite.parentId === site.siteId),
-    );
-
-    if (childSites.length === 0) return null;
-
-    if (isLeafLevel) {
-        return (
-            <ReviewSiteLeafRows
-                sites={childSites}
-                getDisplayName={site => site.name ?? 'Unknown'}
-                sessionCountsBySiteId={sessionCountsBySiteId}
-            />
-        );
-    }
-
-    const locationTypeName = getLocationTypeName(childSites[0]!);
-
-    return (
-        <div className="space-y-1">
-            {childSites.map(site => {
-                const descendantSites = getSiteAndDescendants(
-                    sites,
-                    site.siteId,
-                ).filter(
-                    descendant =>
-                        descendant.siteId !== site.siteId &&
-                        !sites.some(
-                            otherSite =>
-                                otherSite.parentId === descendant.siteId,
-                        ),
-                );
-
-                return (
-                    <ReviewCollapsibleLocationGroup
-                        key={`${parentPath}/${site.name}`}
-                        locationName={site.name ?? 'Unknown'}
-                        locationTypeName={locationTypeName}
-                        sitesInGroup={descendantSites}
-                        parentPath={parentPath}
-                        sessionCountsBySiteId={sessionCountsBySiteId}
-                        expandedSitePaths={expandedSitePaths}
-                        onToggle={onToggle}
-                    >
-                        <HierarchicalSiteHierarchy
-                            sites={sites}
-                            parentSiteId={site.siteId}
-                            parentPath={`${parentPath}/${site.name}`}
-                            sessionCountsBySiteId={sessionCountsBySiteId}
-                            expandedSitePaths={expandedSitePaths}
-                            onToggle={onToggle}
-                        />
-                    </ReviewCollapsibleLocationGroup>
-                );
-            })}
-        </div>
-    );
+function getVisitCoverageBackgroundColor(
+    percentage: number,
+    highThreshold = 80,
+    mediumThreshold = 50,
+): string {
+    if (percentage >= highThreshold) return 'bg-success/10 hover:bg-success/20';
+    if (percentage >= mediumThreshold)
+        return 'bg-warning/10 hover:bg-warning/20';
+    return 'bg-destructive/10 hover:bg-destructive/20';
 }
 
 interface ReviewSiteHierarchyProps {
     sites: Site[];
-    depth: number;
     parentPath: string;
     sessionCountsBySiteId: Map<number, ReviewSiteSessionSummary>;
     expandedSitePaths: Set<string>;
-    onToggle: (path: string) => void;
+    onTogglePath: (path: string, descendantPaths: string[]) => void;
 }
 
 export default function ReviewSiteHierarchy({
     sites,
-    depth,
     parentPath,
     sessionCountsBySiteId,
     expandedSitePaths,
-    onToggle,
+    onTogglePath,
 }: ReviewSiteHierarchyProps) {
-    if (sites.length === 0) return null;
-
-    return isLegacySite(sites[0]!) ? (
-        <LegacySiteHierarchy
+    return (
+        <SiteHierarchy
             sites={sites}
-            depth={depth}
             parentPath={parentPath}
-            sessionCountsBySiteId={sessionCountsBySiteId}
             expandedSitePaths={expandedSitePaths}
-            onToggle={onToggle}
-        />
-    ) : (
-        <HierarchicalSiteHierarchy
-            sites={sites}
-            parentSiteId={undefined}
-            parentPath={parentPath}
-            sessionCountsBySiteId={sessionCountsBySiteId}
-            expandedSitePaths={expandedSitePaths}
-            onToggle={onToggle}
+            onTogglePath={onTogglePath}
+            renderLeafRows={(leafSites, getDisplayName) => (
+                <ReviewSiteLeafRows
+                    sites={leafSites}
+                    getDisplayName={getDisplayName}
+                    sessionCountsBySiteId={sessionCountsBySiteId}
+                />
+            )}
+            renderGroupContent={sitesInGroup => {
+                const visitedCount = sitesInGroup.filter(
+                    site =>
+                        (sessionCountsBySiteId.get(site.siteId)?.sessionCount ??
+                            0) > 0,
+                ).length;
+                const visitedPercentage =
+                    sitesInGroup.length > 0
+                        ? Math.round((visitedCount / sitesInGroup.length) * 100)
+                        : 0;
+                const needsReviewTotal = sitesInGroup.reduce(
+                    (sum, site) =>
+                        sum +
+                        (sessionCountsBySiteId.get(site.siteId)
+                            ?.needsReviewCount ?? 0),
+                    0,
+                );
+                return {
+                    headerClassName:
+                        getVisitCoverageBackgroundColor(visitedPercentage),
+                    summaryContent: (
+                        <>
+                            {needsReviewTotal > 0 && (
+                                <span className="text-destructive text-xs tabular-nums">
+                                    {`${needsReviewTotal} ${needsReviewTotal === 1 ? 'needs' : 'need'} review`}
+                                </span>
+                            )}
+                            <span className="text-muted-foreground text-xs tabular-nums">
+                                {visitedCount} of {sitesInGroup.length} visited
+                            </span>
+                            <ReviewVisitCoverageBadge
+                                visitedPercentage={visitedPercentage}
+                            />
+                        </>
+                    ),
+                };
+            }}
         />
     );
 }
